feat: Add fetchAll method for Parse pointers
